v5.0.2 - Bitchecksum optimized - Now the loop is going precisely 1kHz (we reduced drastically the time of some operations) - Some components updated (PSoC 3.1)

+++ firmware.cydsn/globals.c
@@ -26,4 +26,4 @@ struct st_dev   device;                 // device related variables
 
 uint8 calibration_flag;
 
-uint16 timer_value;
+uint32 timer_value;
